+++ Flow/src/utopia/flow/parse/file/KeptOpenWriter.scala
@@ -5,13 +5,14 @@ import utopia.flow.parse.AutoClose._
 import utopia.flow.parse.file.FileExtensions._
 import utopia.flow.time.Duration
 import utopia.flow.util.logging.Logger
+import utopia.flow.util.result.TryExtensions._
 import utopia.flow.view.immutable.caching.{ClosesAfterIdle, Lazy}
 
 import java.io.{FileOutputStream, OutputStream, OutputStreamWriter, PrintWriter}
 import java.nio.file.Path
 import scala.concurrent.ExecutionContext
 import scala.io.Codec
-import scala.util.Try
+import scala.util.{Failure, Success, Try}
 
 object KeptOpenWriter
 {
@@ -19,15 +20,20 @@ object KeptOpenWriter
 	  * Creates a new writer wrapper
 	  * @param path Path to the file that is modified
 	  * @param keepOpenDuration Duration how long the file / writer is kept open after each write
-	  * @param codec Implicit encoding to use
+	  * @param rethrows Whether errors thrown by the writer functions should be rethrown
+	 *                 (default = false = errors are caught)
+	 * @param logsFailures Whether all encountered writing-failures should be automatically logged using 'logger'
+	 * @param codec Implicit encoding to use
 	  * @param exc Implicit execution context used for closing the writer asynchronously
-	  * @return A new writer wrapper
+	  * @param logger Logging implementation used
+	 * @return A new writer wrapper
 	  */
-	def apply(path: Path, keepOpenDuration: Duration)
+	def apply(path: Path, keepOpenDuration: Duration, rethrows: Boolean = false, logsFailures: Boolean = false)
 	         (implicit codec: Codec, exc: ExecutionContext, scheduler: Scheduler, logger: Logger) =
 	{
 		val existingPathPointer = Lazy { path.createDirectories() }
-		new KeptOpenWriter(keepOpenDuration)(new FileOutputStream(existingPathPointer.value.get.toFile, true))
+		new KeptOpenWriter(keepOpenDuration, rethrows, logsFailures)(
+			new FileOutputStream(existingPathPointer.value.get.toFile, true))
 	}
 }
 
@@ -37,7 +43,8 @@ object KeptOpenWriter
   * @author Mikko Hilpinen
   * @since 24.7.2022, v1.16
   */
-class KeptOpenWriter(keepOpenDuration: Duration)(generate: => OutputStream)
+class KeptOpenWriter(keepOpenDuration: Duration, rethrows: Boolean = false, logsFailures: Boolean = false)
+                    (generate: => OutputStream)
                     (implicit codec: Codec, exc: ExecutionContext, scheduler: Scheduler, log: Logger)
 {
 	// ATTRIBUTES   ----------------------------
@@ -65,9 +72,22 @@ class KeptOpenWriter(keepOpenDuration: Duration)(generate: => OutputStream)
 	
 	/**
 	  * Performs a write operation
-	  * @param f A function that uses a writer (any thrown errors are caught)
+	  * @param f A function that uses a writer
 	  * @tparam A Function result type
-	  * @return Success if writing (including function call) succeeded, failure otherwise
+	  * @return Success if writing succeeded, failure otherwise
 	  */
-	def apply[A](f: PrintWriter => A) = writerP.keepOpenDuring { _.map(f) }
+	def apply[A](f: PrintWriter => A) = {
+		val result = writerP.keepOpenDuring {
+			case Success(writer) =>
+				if (rethrows)
+					Success(f(writer))
+				else
+					Try { f(writer) }
+			
+			case Failure(error) => Failure(error)
+		}
+		if (logsFailures)
+			result.logWithMessage("Failure while writing")
+		result
+	}
 }